requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
attrs==23.2.0
blink1==0.3.1
click==8.1.7
frozenlist==1.4.1
hidapi==0.14.0
idna==3.7
multidict==6.0.5
python-dotenv==1.0.1
setuptools==70.3.0
slack_bolt==1.19.1
slack_sdk==3.31.0
webcolors==24.6.0
yarl==1.9.4
